extract FeaturedThreadBox into FeaturedThreadContent component that houses FeaturedThreadBox and Comments as separate components
